+++ Libplanet/Blocks/Block.cs
@@ -152,16 +152,18 @@ namespace Libplanet.Blocks
 
         /// <summary>
         /// Executes every <see cref="IAction"/> in the
-        /// <see cref="Transactions"/> and gets the result states.
+        /// <see cref="Transactions"/> and gets result states of each step of
+        /// every <see cref="Transaction{T}"/>.
         /// </summary>
         /// <param name="accountStateGetter">An <see cref="AccountStateGetter"/>
         /// delegate to get a previous state.
         /// A <c>null</c> value, which is default, means a constant function
         /// that returns <c>null</c>.</param>
-        /// <returns>The states immediately after <see cref="IAction"/>s in the
-        /// <see cref="Transactions"/> being executed.</returns>
+        /// <returns>Result states of immediately after
+        /// <see cref="IAction"/>s in each <see cref="Transaction{T}"/>
+        /// being executed.</returns>
         [Pure]
-        public IImmutableDictionary<Address, object> EvaluateActions(
+        public IEnumerable<IAccountStateDelta> EvaluateActions(
             AccountStateGetter accountStateGetter = null
         )
         {
@@ -172,9 +174,8 @@ namespace Libplanet.Blocks
             foreach (Transaction<T> tx in Transactions)
             {
                 delta = tx.EvaluateActions(Hash, Index, delta);
+                yield return delta;
             }
-
-            return delta.GetUpdatedStates();
         }
 
         public void Validate()